+++ src/events.rs
@@ -33,37 +33,44 @@ pub struct TableHeadEvent {
     pub mouse_event: MouseEvent,
 }
 
-macro_rules! impl_event_handler {
+macro_rules! impl_default_rc_fn {
     (
         $(#[$meta:meta])*
         $name:ident<$($ty:ident),*>($($arg_name:ident: $arg_ty:ty),*)
+        $(-> $ret_ty:ty)?
+        $({ default $default_return:expr })?
     ) => {
         $(#[$meta])*
         #[derive(Clone)]
-        pub struct $name<$($ty),*>(Rc<dyn Fn($($arg_ty),*)>);
+        pub struct $name<$($ty),*>(Rc<dyn Fn($($arg_ty),*) $(-> $ret_ty)?>);
 
         impl<$($ty),*> Default for $name<$($ty),*> {
             fn default() -> Self {
-                Self(Rc::new(|$(_: $arg_ty),*| {}))
+                #[allow(unused_variables)]
+                Self(Rc::new(|$($arg_name: $arg_ty),*| {
+                    $($default_return)?
+                }))
             }
         }
 
         impl<F, $($ty),*> From<F> for $name<$($ty),*>
-            where F: Fn($($arg_ty),*) + 'static
+            where F: Fn($($arg_ty),*) $(-> $ret_ty)? + 'static
         {
             fn from(f: F) -> Self { Self(Rc::new(f)) }
         }
 
         impl<$($ty),*> $name<$($ty),*> {
-            pub fn run(&self, $($arg_name: $arg_ty),*) {
-                (self.0)($($arg_name),*);
+            pub fn run(&self, $($arg_name: $arg_ty),*) $(-> $ret_ty)? {
+                (self.0)($($arg_name),*)
             }
         }
     }
 }
 
-impl_event_handler!(
+impl_default_rc_fn!(
     /// New type wrapper of a closure that takes a parameter `T`. This allows the event handler props
     /// to be optional while being able to take a simple closure.
     EventHandler<T>(event: T)
 );
+
+pub(crate) use impl_default_rc_fn;