+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## [Unreleased]
+## [0.10.3] - 2024-08-05
 
 ### Features 🚀
 
+- Added i18n support via the `"i18n"` feature which uses `leptos-i18n`. See the `i18n` example for usage.
 - Added row reader to `TableComponent`
+- Added `default_th_sorting_style` to make it easier to write a custom thead cell render component.
 
 ## [0.10.2] - 2024-06-07